Import version 2.06
* This commit corresponds to the pythondialog 2.06 release, with the
exception that all non-ASCII files (namely INSTALL, README and TODO)
have been recoded from ISO 8859-1 to UTF-8.

* dialog.py: fix a bug with the default_item "common argument"
(corresponding to dialog's --default-item option) thanks to Peter
Mathiasson.

* demo.py: make sure the directory passed to --fselect ends with
os.sep so that its contents can be seen right away in the file
selection box displayed by dialog.

* dialog.py: fix a bug (the standard output of the process running
dialog used to be connected to a pipe) that rendered pythondialog
unusable with recent versions of dialog (thanks, Peter Åstrand!).

* dialog.py: the generic exception is now 'error' ('dialog.error',
if you understand this better). The old name is still there for
backward-compatibility.
* Several new exceptions have appeared: PythonDialogSystemError,
PythonDialogIOError, PythonDialogOSError,
PythonDialogErrorBeforeExecInChildProcess,
PythonDialogReModuleError, UnableToCreateTemporaryDirectory,
PythonDialogBug, ProbablyPythonBug.
* dialog.py: the ExecutableNotFound exception was not mentioned in
the docstring module. Fixed.
* dialog.py: the complete_message() of 'error' instances does not
write a dot at the end of the message anymore.
* dialog.py: rename the ExceptionPrettyIdentifier attribute of
'error' subclasses to ExceptionShortDescription.
* dialog.py: the constructor of 'error' instances can now be
called with no argument if no useful message can be added to the
ExceptionShortDescription.

* dialog.py: added the new DIALOG_EXTRA and DIALOG_HELP return
codes that appeared somewhere between versions 0.9a-20020309a and
0.9b-20040301 of dialog.

* dialog.py: Remove the possibility to choose (from Dialog's
constructor) the values of the DIALOG_{OK,CANCEL,ERROR,...}
environment variables passed to dialog because this is absolutely
useless as far as I can see and clutters the API.

* dialog.py: Dialog's constructor has a new 'compat' parameter
that can be used to enable a compatibility mode with dialog-like
programs whose interface is only slightly different from that of
dialog. The demo runs fine with Xdialog 2.0.6 in the "Xdialog"
compatibility mode. However, I don't want the special cases to
expand too much, so it would be really better for you to report
bugs if your dialog-like program is not dialog-compatible!

* dialog.py: don't try any PATH components (nor any components
from ":/bin:/usr/bin" if PATH is undefined) if the 'dialog'
parameter of Dialog's constructor contains a '/'.

* dialog.py: somewhere between versions 0.9a-20020309a and
0.9b-20040301, dialog started not to quote tags in the output of
--checklist if they didn't contain any space. Adapt the dialog
invocation (now using --separate-output) and the output parsing
accordingly (still works with 0.9a-20020309a).

* dialog.py: the data fed to dialog in Dialog.gauge_update()
happened to work with dialog 0.9a-20020309a but did not conform to
the manual page and presumably broke Xdialog. This is fixed,
thanks to Peter Åstrand.

* dialog.py: modifications in Dialog.scrollbox():
    - don't use the old insecure tempfile.mktemp() anymore
      (tempfile.mkstemp() didn't exist before Python 2.3)
    - don't display a title for the box if no title was in kwargs
    - the return value is now that of the dialog-like program
    - UnableToCreateTemporaryDirectory is now raised if for some
      strange reason, we cannot create a temporary directory

* dialog.py: review every function to catch possible exceptions
such as IOError and OSError, in order to turn them into subclasses
of 'error' (such as PythonDialogIOError and PythonDialogOSError).
Rare beasts such as MemoryError are still not caught. Such an
enterprise would be unreasonable, if not simply impossible.
* dialog.py: update the various docstrings to show which
exceptions every function can raise. To make this manageable, many
functions refer to the docstrings of internal, heavily-used
functions such as Dialog.__perform whose docstrings are,
unfortunately but rightfully, not included in the HTML
documentation generated by pydoc. Some sort of automated
documentation generation system would be needed to solve this
problem in a satisfactory way.

* demo.py: minor update with respect to Debian sarge's release
date forecasts...