Add code to deal with Xdialog's incompatibilities
* This change is part of pythondialog 2.7 (uploaded to SourceForge by
Peter Åstrand in 2004), but was not included in version 2.08 from
Florent Rougon, the latter being based on version 2.06 under the
assumption that no one was using Xdialog anymore (since it required the
long-discontinued GTK+ 1 toolkit). It is now included in order to make
everyone happy.

* dialog.py: add Peter Åstrand's modifications to deal with
Xdialog's incompatibilities with respect to dialog:
  - new "use_stdout" keyword argument for Dialog.__init__()
  - factoring of final newline removals with the addition of
    Dialog._strip_xdialog_newline().