Import version 2.7
* This commit corresponds to the pythondialog 2.7 release as
distributed at <http://sourceforge.net/projects/pythondialog/>, with the
following exceptions:
  - it does not remove the ChangeLog file, although the 2.7 release has
    no such file; the main reason for this is to avoid polluting the
    repository, and especially diffs between version 2.7 and later
    releases, with a non-code file disappearing in one release and
    reappearing in the next.
  - all non-ASCII files (namely INSTALL, README and TODO) have been
    recoded from ISO 8859-1 to UTF-8, for the same reason of not
    polluting diffs between releases (or revisions).

* dialog.py: add a 'use_stdout' optional argument to the Dialog
constructor. When set to True, all calls to the dialog-like
backend include the --stdout option, and pythondialog accordingly
expects said backend to report on stdout instead of stderr. This
is useless with dialog, but reportedly necessary with Xdialog;
therefore, it is only enabled by default when passing
compat="Xdialog" to the Dialog constructor.

* dialog.py: rename the following Dialog methods and attributes to
start with only one underscore: __dialog_prg, __call_program(),
__wait_for_program_termination(), __perform(),
__perform_no_options() and __gauge_process().

* dialog.py (Dialog._strip_xdialog_newline): new method to factor
out the removal of the final newline in Xdialog-compatibility mode.

* Set "Peter Astrand" as maintainer and among authors in setup.py
and AUTHORS, home page to "http://pythondialog.sourceforge.net/",
document the requirement of Python 2.2 or newer in INSTALL, remove
ChangeLog from MANIFEST.in (???) and bump version to 2.7.