Import version 2.09
* dialog.py: new supported widgets: editbox, inputmenu, mixedform,
mixedgauge, pause, passwordform, progressbox.

* demo.py: general improvements, such as:
    - hopefully more helpful dialog when the user fails to select
      a file in fselect;
    - since actually selecting a file with this widget is already
      boring after the second time, the widget can be exited by
      pressing Esc or the Cancel button (in which case the parts
      of the demo that need the file path will be skipped).
    - replace a few calls to Dialog.scrollbox() by calls to
      Dialog.msgbox(), since the latter provides nice automatic
      line wrapping.

* demo.py: support GNU-style option passing with getopt.py; you
can use --help to get a list of available options
* demo.py: no need to change a global variable anymore to switch
the demo to "fast mode", just use --fast
* demo.py: add --test-suite mode, mainly for developers: it
tests *all* widgets, not only those included in the "default
mode", and automatically enables "fast mode".

* dialog.py (_call_program): make the function more generic. stdin
redirection doesn't involve automatic pipe(2) creation in
_call_program() anymore; instead, _call_program expects a file
descriptor when it is asked to redirect dialog's stdin (parameter
'redir_child_stdin_from_fd').

The caller may still decide to create a pipe and pass its file
descriptor for reading as the 'redir_child_stdin_from_fd'
parameter, but the new possibility of redirecting dialog's stdin
from an arbitrary file descriptor allows for instance to redirect
it from an existing file, network socket... This is used to
implement --progressbox cleanly.

This change has the additional benefit of simplifying the API,
since _call_program()'s return value is always a 2-element tuple
now.

* dialog.py (_call_program): new close_fds option causing the
child process to close the specified file descriptors before the
execve(2) system call. This is useful for instance to have the
child close an end of a pipe he isn't going to use. Without that,
deadlocks could happen because of the child never seeing EOF from
the pipe.

* dialog.py: use warnings.warn(..., DeprecationWarning) for
obsolete functions.

* dialog.py: remove convoluted syntax *(<list>,) that was used at
several places.

* dialog.py: prefix attributes for internal use (such as
Dialog._call_program) with a single underscore instead of a double
one: we don't need the name mangling here. These underscores in
dialog.py are just an indication that the attribute is "internal"
and thus subject to API changes, etc. Thanks to Peter Åstrand for
pointing this out.

* setup.py:
  - improve the long description, use ReStructuredText
  - add Trove classifiers
  - add download_url

* Review and update README, TODO...

* dialog.py: add support for --dselect
* dialog.py: add support for DIALOG_ITEM_HELP
* demo.py: small fixes