Escape non-option arguments to dialog that start with --
* dialog.py (_dash_escape): new function that inserts a '--' element
before every element in a list that starts with '--' (returns a new
list).

* dialog.py (_dash_escape_nf): same as _dash_escape, but ignores the
first element.

* dialog.py (Dialog.dash_escape, Dialog.dash_escape_nf): new class
methods for public use, for those using Dialog.add_persistent_args().

* dialog.py (Dialog._call_program): new keyword argument 'dash_escape'
to control the dash-escaping of the argument list; by default, the
appropriate part is filtered through _dash_escape_nf().

* dialog.py (Dialog._perform): new keyword argument 'dash_escape' that
is passed as is to Dialog._call_program() (closes: SF#5; thanks to
Denilson Figueiredo de Sá for the report).

* dialog.py: use _dash_escape_nf() where appropriate in
_common_args_syntax for options that are passed as keyword arguments
(e.g., title="...").

* dialog.py (Dialog.setBackgroundTitle): use self.dash_escape_nf() in
this long-deprecated method.