New method Dialog.set_background_title(), better Dialog doc
* dialog.py (Dialog.set_background_title): new method. Now that we deal
with dash escaping, such a method does become useful since it avoids
users having to use the low-level Dialog.add_persistent_args() and
bother with dash escaping.

* dialog.py (Dialog.setBackgroundTitle): this old method from
pythondialog 1.0 is still deprecated because its case style is
inconsistent with the rest of pythondialog; it simply calls
Dialog.set_background_title() after printing a DeprecationWarning.

* dialog.py (Dialog): add set_background_title, maxsize and
backend_version to the docstring and reorganize a bit the list of public
methods.

* demo.py: use the new method.