Misc demo improvements
* demo.py: move a bunch of widget demos from additional_widgets() to the
main demo() function to give them more visibility. The remaining widgets
in additional_widgets() either have little warts (like programbox which
is waiting for a fix in the dialog backend), are cumbersome to use in
the demo (this is the case of progressbox_demo_with_filepath), or are
almost identical to widgets already presented in the main part of the
demo.

* demo.py: use the current local time to initialize the timebox widget.

* demo.py: minor improvements.