Add an 'is_widget' attribute to Dialog widget-producing methods
* dialog.py (widget): new decorator to mark the Dialog methods that
provide a widget. As explained in the docstring, this allows code to
perform automatic operations on these specific methods. For instance,
one can define a class that behaves similarly to Dialog, except that
after every widget-producing call, it spawns a "confirm quit" dialog if
the widget returned DIALOG_ESC, and loops in case the user doesn't
actually want to quit.