Improve structure and ESC handling in the demo
* demo.py: new MyApp class that implements the core of the demo. This
class relies on a new MyDialog class that automatically wraps every
widget-producing method of dialog.Dialog in order to display the
"confirm quit" dialog if the user presses the Escape key or the Cancel
button. This class also provides a few dialog-related methods used in
the demo.

* demo.py: This new structure should completely fix handling of the
Escape key, which was not satisfactory in previous versions since it
required a while loop for every widget call that made the code redundant
and harder to read. The new wrapping mechanism is completely transparent
for most of the code in MyApp, which thus becomes shorter, more reliable
and easier to read. The "magic" is contained within the MyDialog class.

* A sample program for newcomers to pythondialog with absolutely no
magic will be added in the next commit (simple_example.py).