Better reporting of dialog errors
* dialog.py: when dialog exits with status DIALOG_ERROR, write its
output as part of the DialogError exception that is raised. This should
make it much easier to understand the cause of errors. This requires
reading the dialog output before wait()ing for it to exit, which is a
good thing in any case, as big amounts of output could cause a kind of
deadlock, since dialog would be blocked with its output pipe full while
we would be wait()ing for it to exit.