Full help support for all widgets
* dialog.py: add support for --help-status and --help-tags.

* dialog.py: review all widgets, making sure that all help-related
options can effectively be used (in particular 'item_help' and
'help_status') and that the relevant output is returned to the caller in
a suitable form.

* dialog.py: compatibility should not be affected for 'menu' nor for
basic usage when only the "help" exit code is checked, without parsing
the output. With this commit, widget-producing methods do check if the
exit code is "help" and, in that case, parse the output in order to
return a useful and easy to use Python object to the caller. This means
that "HELP " prefixes are removed, shell-style quoting and
backslash-escaping is undone, and that the output of --help-status is
returned as a hopefully useful data structure.

* demo.py: add help support to many widgets, showing how to use the new
facilities.