Remove _create_temporary_directory() in favor of tempfile.NamedTemporaryFile()
* dialog.py (Dialog.scrollbox): use tempfile.NamedTemporaryFile()
instead of our custom _create_temporary_directory() function.

* There was no security problem in _create_temporary_directory() as far
as I know, however it is usually better to use well-tested library
functions when possible instead of custom ones. When
_create_temporary_directory() was written, what was available from the
tempfile module was not satisfactory, but this is not the case anymore.

* dialog.py: BACKWARD INCOMPATIBILITY: the
UnableToCreateTemporaryDirectory exception is not defined anymore.
Dialog.scrollbox() now creates a temporary file without any temporary
directory, therefore there is no place anymore for this exception to be
used. The equivalent condition in tempfile.NamedTemporaryFile()
generates an OSError exception (more precisely, a FileExistsError in
Python 3.3 or later, which is a subclass of OSError). As usual, this
exception is wrapped by pythondialog and seen as a PythonDialogOSError
by user code.

* As a conclusion, wherever user code was expecting
UnableToCreateTemporaryDirectory in previous versions, it should now
expect a PythonDialogOSError, consistently with the tempfile module and
OSError wrapping by pythondialog.