New 'features' concept
* dialog.py: new concept called 'features' that is similar to Python's
__future__ statement. The general idea is that users can globally modify
pythondialog's behavior by calling:

  dialog.enable_feature(FEATURE)

right after importing the 'dialog' module, in order to ensure consistent
behavior. In this call, FEATURE must be a member of the 'dialog.Feature'
enum. Using this mechanism therefore requires a Python installation
containing the 'enum' module. Since it is part of the standard library
in Python 3.4, this should not be a significant problem.

Note: the Python 3.4 'enum' module has been backported to older Python
      versions under the name 'enum34'.

* dialog.py: this commit adds the following module-level variables:

   _HAS_ENUM, _features

as well as the Feature class (which is an enum).

* dialog.py (enable_feature, disable_feature, feature_enabled_p): new
functions.