New 'autowidgetsize' feature
* dialog.py: add the 'autowidgetsize' feature. When enabled,
pythondialog's widget-producing methods behave as if width=0, height=0,
etc. had been passed, except where these parameters are explicitely
specified with different values. This feature is currently marked as
experimental, please give some feedback.

Note: in order to differentiate between a default value obtained when
'autowidgetsize' is disabled and an explicitely-specified width, height,
etc., the size parameters modified by this commit now default to None.
In order to compensate for this information loss, the effective default
values when 'autowidgetsize' is disabled are now mentioned in the
docstrings of the corresponding widget-producing methods.

* dialog.py (Dialog._default_size): new method returning the appropriate
default values for the size parameters depending on whether
'autowidgetsize' is enabled or not. This allows tight factoring of the
code that had to be added to almost every widget.