demo.py: programbox and progressboxoid changes
* demo.py (MyApp.programbox_demo): remove the \n character at the end of
'text'. Otherwise, dialog will see *two* newline chars in a row at the
end and will faithfully print a blank line. This, in turn, may cause
unwanted scrolling if the window doesn't have enough space for that
line.

* demo.py: run programbox_demo from MyApp.demo() if the dialog version
is >= 1.2-20140112, otherwise leave it in MyApp.additional_widgets().
Versions older than 1.2-20140112 have a bug that makes it unsuitable
for the main demo.

* demo.py (MyApp.progressboxoid): allow additional arguments to be
passed as **kwargs.