Miscellaneous little changes to ancillary files
* MANIFEST.in: remove "global-exclude *~" and
"global-exclude .gitattributes" because they cause warnings upon
installation with pip that may frighten users, and only provide a safety
net against hypothetical bugs that don't exist in the current version.

* TODO: update the file now that the 'autowidgetsize' option has been
implemented.

* setup.py: don't print the "No .git directory, using the 'ChangeLog'
file as is" message anymore, as it pollutes the installation with pip,
possibly causing users to fear that something went wrong.