Change 'autowidgetsize' into a Dialog constructor option
* dialog.py: remove the 'features' mechanism and change 'autowidgetsize'
into a Dialog constructor option (boolean, keyword-only argument). This
is simpler, doesn't require any recent package and makes it possible to
have several Dialog instances with different values for the
'autowidgetsize' setting in the same process... The 'features' mechanism
is removed because it will have no use after this change.

* examples/with-autowidgetsize/demo.py,
  examples/with-autowidgetsize/simple_example.py: adapt to this change.