Fix calendar default values and docstring
* dialog.py (Dialog.calendar): change the default values for day, month
and year from 0 to -1. This is unfortunately BACKWARD-INCOMPATIBLE,
but 0 doesn't work well for day and month in dialog, and it would be
quite unintuitive to have them default to -1 with the year still
defaulting to 0.

* dialog.py (Dialog.calendar): improve the docstring clarity.

* examples/demo.py (calendar_demo_with_help): use -1 for the initial
values of day, month and year. This is only for clarity; the previous
code would work just as well.