Remove current directory from the default used when PATH is unset
* dialog.py (_find_in_path): in case PATH was unset, the default value
used to start with a colon, which caused pythondialog to look for the
executable in the current directory if (1) the specified executable
didn't contain a '/' and (2) the PATH environment variable was unset.
This behavior was conformant to what glibc's execvp() function did until
version 2.24 from 2016, but wasn't ideal (if a pythondialog-based
program had its current directory world-writable, this could be used to
trap users whose PATH is unset---which is rather unlikely, fortunately).
This suboptimal default behavior when PATH is unset was changed in glibc
2.24, and we are following suit here, i.e., the new default is
"/bin:/usr/bin". Also clarify the docstring.

* dialog.py (_path_to_executable): return the result after applying
os.path.realpath() and clarify the docstring. Using os.path.realpath()
at this point ensures that the path to the dialog-like executable is
resolved at the time the Dialog instance is created. This way, a
pythondialog-based program can change its current directory after
creating a Dialog instance without fearing that this might cause
subsequent pythondialog calls to fail or to invoke an executable from a
different directory.

* dialog.py (Dialog.__init__): adapt the docstring to reflect the
aforementioned change when PATH is unset (as mentioned above for
_find_in_path()).