Fix related to the 'timeout' option (--timeout for dialog)
* dialog.py (Dialog._perform): ignore the dialog output when a timeout
occurred, which we know thanks to the exit code. Without this change,
output parsing or checking (e.g., in Dialog._widget_with_no_output())
would be broken by the '\ntimeout\n' string printed by dialog in such
cases (seen with msgbox, at least).

It may be that parsing is still disturbed for some widgets when a
timeout occurs, if they never produce an empty output in a no-timeout
situation: this remains to be checked.